Extract .syncignore-to-rsync translation into syncignore.py
The pattern helpers were pure functions with no Folder state, and the
class had grown to twice the size of any other module. The new module
follows the repo's caller-before-callee ordering, with the public
get_rsync_filter_rules() on top; Folder._get_rsync_filters() keeps the
shell quoting, since assembling the command line is its job.
