+++ livesync/folder.py
@@ -11,6 +11,7 @@ import pathspec
 import watchfiles
 
 from .run_subprocess import run_subprocess
+from .syncignore import get_rsync_filter_rules
 
 
 class Folder:
@@ -66,92 +67,9 @@ class Folder:
                 ignores.append(pattern)
         return ignores
 
-    @staticmethod
-    def _is_directory_pattern(pattern: str) -> bool:
-        return pattern.endswith('/') or pattern.endswith('\\')
-
-    @staticmethod
-    def _is_anchored(pattern: str) -> bool:
-        """Tell whether gitignore anchors this pattern to the source folder.
-
-        A slash at the start or in the middle anchors the pattern, while a slashless one like
-        ``node_modules/`` matches at any depth. rsync patterns anchor only on a leading slash,
-        so the middle-slash case needs one adding.
-        """
-        return '/' in pattern.rstrip('/\\')
-
-    @classmethod
-    def _anchor(cls, pattern: str) -> str:
-        return pattern if pattern.startswith('/') or not cls._is_anchored(pattern) else f'/{pattern}'
-
-    @staticmethod
-    def _without_trailing_slash(pattern: str) -> str:
-        return pattern.rstrip('/\\')
-
-    @classmethod
-    def _normalise_for_descendant_check(cls, pattern: str) -> str:
-        return cls._without_trailing_slash(pattern.lstrip('!').lstrip('/'))
-
-    @classmethod
-    def _directory_pattern_has_negation(cls, pattern: str, negations: List[str]) -> bool:
-        directory = cls._normalise_for_descendant_check(pattern)
-        if not directory or not negations:
-            return False
-        if any(char in directory for char in '*?['):
-            return True  # a glob may cover a negated path, so keep the directory traversable
-        for negation in negations:
-            candidate = cls._normalise_for_descendant_check(negation)
-            if candidate == directory or candidate.startswith(f'{directory}/'):
-                return True
-            if not cls._is_anchored(pattern) and f'/{directory}/' in candidate:
-                return True  # a slashless pattern matches at any depth, so a nested negation counts
-        return False
-
-    @staticmethod
-    def _ancestor_directory_patterns(pattern: str) -> List[str]:
-        parts = [part for part in pattern.strip('/\\').split('/')[:-1] if part]
-        return ['/' + '/'.join(parts[:index]) + '/' for index in range(1, len(parts) + 1)]
-
-    @classmethod
-    def _rsync_rules_for_negation(cls, pattern: str) -> List[str]:
-        rule = cls._anchor(pattern[1:])
-        rules = [f'+ {ancestor}' for ancestor in cls._ancestor_directory_patterns(rule)]
-        if cls._is_directory_pattern(rule):
-            rules.append(f'+ {rule}')
-            rules.append(f'+ {rule}**')
-        else:
-            rules.append(f'+ {rule}')
-        return rules
-
-    def _get_rsync_filter_rules(self) -> List[str]:
-        negations = [pattern for pattern in self._ignores if pattern.startswith('!')]
-        rules: List[str] = []
-        for pattern in reversed(self._ignores):
-            if pattern.startswith('!'):
-                rules.extend(self._rsync_rules_for_negation(pattern))
-            elif self._directory_pattern_has_negation(pattern, negations):
-                # Ancestor '+' rules reopen this directory for traversal, so pruning it is not
-                # an option: exclude its contents at any depth instead.
-                anchored = self._anchor(pattern)
-                rules.append(f'- {self._without_trailing_slash(anchored)}/**')
-                if not self._is_directory_pattern(pattern):
-                    rules.append(f'- {anchored}')
-            else:
-                rules.append(f'- {self._anchor(pattern)}')
-
-        deduplicated: List[str] = []
-        for rule in rules:
-            if rule not in deduplicated:
-                deduplicated.append(rule)
-        return deduplicated
-
     def _get_rsync_filters(self) -> str:
-        """Convert .syncignore patterns to rsync filter rules, supporting negation (!) prefixes.
-
-        Rules are emitted in reverse .syncignore order so rsync's first-match-wins behavior
-        matches pathspec's last-match-wins behavior.
-        """
-        return ''.join(f' --filter={shlex.quote(rule)}' for rule in self._get_rsync_filter_rules())
+        """Convert .syncignore patterns to rsync filter rules, supporting negation (!) prefixes."""
+        return ''.join(f' --filter={shlex.quote(rule)}' for rule in get_rsync_filter_rules(self._ignores))
 
     def _is_ignored(self, filepath: str) -> bool:
         path = Path(filepath)
